Make `verify` & some shared methods default
