+++ crates/sncast/src/starknet_commands/verification/base.rs
@@ -4,43 +4,31 @@ use camino::Utf8PathBuf;
 use reqwest::StatusCode;
 use serde::Serialize;
 use sncast::response::structs::VerifyResponse;
-use sncast::Network;
 use starknet::core::types::Felt;
 use std::ffi::OsStr;
 use walkdir::WalkDir;
 
 #[async_trait]
 pub trait VerificationInterface {
-    fn new(network: Network, workspace_dir: Utf8PathBuf) -> Self;
-    async fn verify(
-        &self,
-        contract_address: Option<Felt>,
-        class_hash: Option<Felt>,
-        class_name: String,
-    ) -> Result<VerifyResponse>;
+    fn get_workspace_dir(&self) -> Utf8PathBuf;
+
     fn gen_explorer_url(&self) -> String;
-}
 
-pub struct BaseVerificationInterface {
-    pub network: Network,
-    pub workspace_dir: Utf8PathBuf,
-}
-
-impl BaseVerificationInterface {
-    pub fn read_workspace_files(&self) -> Result<serde_json::Map<String, serde_json::Value>> {
+    fn read_workspace_files(&self) -> Result<serde_json::Map<String, serde_json::Value>> {
         // Read all files name along with their contents in a JSON format
         // in the workspace dir recursively
         // key is the file name and value is the file content
         let mut file_data = serde_json::Map::new();
 
         // Recursively read files and their contents in workspace directory
-        for entry in WalkDir::new(self.workspace_dir.clone()).follow_links(true) {
+        let workspace_dir = self.get_workspace_dir();
+        for entry in WalkDir::new(workspace_dir.clone()).follow_links(true) {
             let entry = entry?;
             let path = entry.path();
             if path.is_file() {
                 if let Some(extension) = path.extension() {
                     if extension == OsStr::new("cairo") || extension == OsStr::new("toml") {
-                        let relative_path = path.strip_prefix(self.workspace_dir.clone())?;
+                        let relative_path = path.strip_prefix(workspace_dir.clone())?;
                         let file_content = std::fs::read_to_string(path)?;
                         file_data.insert(
                             relative_path.to_string_lossy().into_owned(),
@@ -53,7 +41,7 @@ impl BaseVerificationInterface {
         Ok(file_data)
     }
 
-    pub async fn send_verification_request(
+    async fn send_verification_request(
         &self,
         url: String,
         payload: VerificationPayload,
@@ -78,6 +66,24 @@ impl BaseVerificationInterface {
             let message = api_res.text().await.context("Failed to verify contract")?;
             Err(anyhow!(message))
         }
+    }
+
+    async fn verify(
+        &self,
+        contract_address: Option<Felt>,
+        class_hash: Option<Felt>,
+        class_name: String,
+    ) -> Result<VerifyResponse> {
+        let file_data = self.read_workspace_files()?;
+        let source_code = serde_json::Value::Object(file_data);
+        let payload = VerificationPayload {
+            class_name,
+            contract_address,
+            class_hash,
+            source_code,
+        };
+        let url = self.gen_explorer_url();
+        self.send_verification_request(url, payload).await
     }
 }
